move pluginlib in its own folder (port 83 to ros2 branch) (#95)
* move pluginlib in its own folder (port 83 to ros2 branch)

* fix most linter errors

* forward port of #88

Continue loading classes on error
