+++ pluginlib/CMakeLists.txt
@@ -58,7 +58,7 @@ install(
 #   endif()
 
 #   include(CheckCXXCompilerFlag)
-#   CHECK_CXX_COMPILER_FLAG("-std=c++11" COMPILER_SUPPORTS_CXX11)
+#   check_cxx_compiler_flag("-std=c++11" COMPILER_SUPPORTS_CXX11)
 #   if(COMPILER_SUPPORTS_CXX11)
 #     catkin_add_gtest(${PROJECT_NAME}_unique_ptr_test test/unique_ptr_test.cpp)
 #     if(TARGET ${PROJECT_NAME}_unique_ptr_test)
