+++ pluginlib/include/pluginlib/class_loader_base.hpp
