+++ pluginlib/include/pluginlib/class_loader_imp.hpp
@@ -316,7 +316,13 @@ std::map<std::string, ClassDesc> ClassLoader<T>::determineAvailableClasses(
   for (std::vector<std::string>::const_iterator it = plugin_xml_paths.begin();
     it != plugin_xml_paths.end(); ++it)
   {
-    processSingleXMLPluginFile(*it, classes_available);
+    try {
+      processSingleXMLPluginFile(*it, classes_available);
+    } catch (const pluginlib::InvalidXMLException & e) {
+      RCUTILS_LOG_ERROR_NAMED("pluginlib.ClassLoader",
+        "Skipped loading plugin with error: %s.",
+        e.what());
+    }
   }
 
   RCUTILS_LOG_DEBUG_NAMED("pluginlib.ClassLoader", "Exiting determineAvailableClasses()...");
@@ -670,15 +676,15 @@ void ClassLoader<T>::processSingleXMLPluginFile(
   tinyxml2::XMLElement * config = document.RootElement();
   if (NULL == config) {
     throw pluginlib::InvalidXMLException(
-            "XML Document has no Root Element. "
-            "This likely means the XML is malformed or missing.");
+            "XML Document '" + xml_file +
+            "' has no Root Element. This likely means the XML is malformed or missing.");
     return;
   }
   if (!(strcmp(config->Value(), "library") == 0 ||
     strcmp(config->Value(), "class_libraries") == 0))
   {
     throw pluginlib::InvalidXMLException(
-            "The XML document given to add must have either \"library\" or "
+            "The XML document '" + xml_file + "' given to add must have either \"library\" or "
             "\"class_libraries\" as the root tag");
     return;
   }
