+++ pluginlib/include/pluginlib/exceptions.hpp
@@ -47,7 +47,7 @@ public:
   : std::runtime_error(error_desc) {}
 };
 
-/// An exception class thrown when pluginlib is unable to load a plugin XML file.
+/// Thrown when pluginlib is unable to load a plugin XML file.
 /**
  * \class InvalidXMLException
  */
@@ -58,7 +58,7 @@ public:
   : PluginlibException(error_desc) {}
 };
 
-/// An exception class thrown when pluginlib is unable to load the library associated with a given plugin.
+/// Thrown when pluginlib is unable to load the library associated with a given plugin.
 /**
  * \class LibraryLoadException
  */
@@ -69,7 +69,7 @@ public:
   : PluginlibException(error_desc) {}
 };
 
-/// An exception class thrown when pluginlib is unable to instantiate a class loader.
+/// Thrown when pluginlib is unable to instantiate a class loader.
 /**
  * \class ClassLoaderException
  */
@@ -80,7 +80,7 @@ public:
   : PluginlibException(error_desc) {}
 };
 
-/// An exception class thrown when pluginlib is unable to unload the library associated with a given plugin.
+/// Thrown when pluginlib is unable to unload the library associated with a given plugin.
 /**
  * \class LibraryUnloadException
  */
@@ -91,7 +91,7 @@ public:
   : PluginlibException(error_desc) {}
 };
 
-/// An exception class thrown when pluginlib is unable to create the class associated with a given plugin.
+/// Thrown when pluginlib is unable to create the class associated with a given plugin.
 /**
  * \class CreateClassException
  */
