+++ pluginlib/test/unique_ptr_test.cpp
